Fix trivial compile error in thread_perftest.cc
Win_x64 LKGR builder has been red for awhile due to a trivial
compile-time conversion error (size_t to int).

BUG=
TEST=run base_perftest on linux
TBR=thakis@chromium.org

Review URL: https://codereview.chromium.org/203543009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@258145 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/threading/thread_perftest.cc
@@ -158,7 +158,7 @@ class EventPerfTest : public ThreadPerfTest {
   virtual void Reset() OVERRIDE { events_.clear(); }
 
   void WaitAndSignalOnThread(size_t event) {
-    int next_event = (event + 1) % events_.size();
+    size_t next_event = (event + 1) % events_.size();
     int my_hops = 0;
     do {
       events_[event]->Wait();
